fix: modify lead_time_minutes to bigint

+++ plugins/github/tasks/issue_convertor.go
@@ -85,7 +85,7 @@ func ConvertIssues(taskCtx core.SubTaskContext) errors.Error {
 				AssigneeName:    issue.AssigneeName,
 				CreatorId:       accountIdGen.Generate(data.Options.ConnectionId, issue.AuthorId),
 				CreatorName:     issue.AuthorName,
-				LeadTimeMinutes: issue.LeadTimeMinutes,
+				LeadTimeMinutes: int64(issue.LeadTimeMinutes),
 				Url:             issue.Url,
 				CreatedDate:     &issue.GithubCreatedAt,
 				UpdatedDate:     &issue.GithubUpdatedAt,
